model-api-functions/requirements.txt
azure-functions
scipy>=1.2.0
numpy>=1.16.0
scikit-learn==0.20.3